fix(vaults): refactor abis
