+++ services/vaults/lib/vaults/v1.js
@@ -2,7 +2,7 @@
 
 const Web3 = require('web3');
 const yRegistryAbi = require('../../../../abi/yRegistry.json');
-const vaultAbi = require('../../../../abi/vaultV1.json');
+const vaultAbi = require('../../../../abi/vaults/yvault_v1.json');
 
 const yRegistryAddress = '0x3ee41c098f9666ed2ea246f4d2558010e59d63a0';
 const web3 = new Web3(process.env.WEB3_ENDPOINT);
